deploy/requirements.txt
flask
gunicorn
scikit-learn==1.0.2
google-cloud-storage

